+++ src/physical_plan.rs
@@ -241,7 +241,7 @@ impl PhysicalOperator for PhysicalVectorSearch<'_> {
         if let Some(vec) = vector {
             let neighbors = {
                 let index = self.storage.hnsw.load();
-                let vs = self.storage.vector_store.read();
+                let vs = self.storage.vector_store[0].read();
                 index.search_nearest(&vec, None, None, &crate::node::ALL_BITSET, 5, Some(&vs))
             };
             for (id, score) in neighbors {